fix(cli): move headlessRender + scriptLoader under src/ to satisfy tsconfig.cli rootDir
The CLI build (tsconfig.cli.json) sets rootDir=./src and includes only
src/. headlessRender lived in scripts/lib/ and importing it from src/cli/
crashed tsc with TS6059 (file not under rootDir), which cascaded into
~30 false-positive type errors on the components/ tree.

Move:
  scripts/lib/headlessRender.ts → src/render/headlessRender.ts
  scripts/lib/scriptLoader.ts   → src/script-runtime/scriptLoader.ts
                                 + thin re-export shim at the old path so
                                   captureDemo and existing tests keep working

The RenderView type that DemoPlayerPage and headlessRender both need
moves into src/render/views.ts — a tiny boundary file that's safe for
the browser bundle (no playwright/sharp imports) and the CLI build alike.
