+++ scripts/lib/scriptLoader.ts
@@ -1,23 +1,4 @@
-import { readFileSync } from 'node:fs';
-import { basename } from 'node:path';
-import { runScript } from '../../src/script-runtime/runScript';
-import type { FeatureRecord } from '../../src/intent/featureRecord';
-import type { FeatureKind } from '../../src/intent/types';
-import type { ParamTable } from '../../src/runtime/paramTable';
-
-export interface LoadedScript {
-  source: string;
-  features: { id: string; kind: FeatureKind; record: FeatureRecord }[];
-  paramTable: ParamTable;
-}
-
-export async function loadScriptFeatures(scriptPath: string): Promise<LoadedScript> {
-  const source = readFileSync(scriptPath, 'utf8');
-  const fileName = basename(scriptPath);
-  const { records, paramTable } = await runScript({ code: source, fileName });
-  return {
-    source,
-    paramTable,
-    features: records.map((r) => ({ id: r.id, kind: r.kind, record: r })),
-  };
-}
+// Re-export so scripts/captureDemo.ts and tests can keep their existing
+// import path. The canonical implementation lives in src/script-runtime/
+// where the CLI build can pick it up via tsconfig.cli.json's rootDir.
+export { loadScriptFeatures, type LoadedScript } from '../../src/script-runtime/scriptLoader';
